Further factorising and added P-val starting condition, but very basic
